requirements.txt
discord.py>=2.3.2

spotipy>=2.25.0
pytest-asyncio>=0.23.0
yt-dlp>=2025.5.22
